backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.7
python-dotenv>=1.0.0
pandas>=1.3.0
numpy>=1.20.0
openpyxl>=3.0.0
pymongo>=4.0.0
dnspython>=2.0.0
google-genai>=1.0.0
requests>=2.25.0
tenacity>=8.0.0
pydantic>=2.0.0
tqdm>=4.66.0
